Reduce duplication in Claim Create Forms
